fix: data type

+++ src/adapters/database/models/balance.py
@@ -1,4 +1,5 @@
 from datetime import UTC, datetime
+from decimal import Decimal
 from uuid import UUID
 
 from sqlmodel import Field, Relationship
@@ -11,9 +12,9 @@ class BalanceModel(BaseModel, table=True):
 
 	__tablename__ = 'balance'
 
-	amount: float = Field(nullable=False)
+	amount: Decimal = Field(nullable=False)
 	user_id: UUID = Field(nullable=False, foreign_key='user.id')
-	user: 'UserModel' = Relationship(back_populates='balance')
+	user: 'UserModel' = Relationship(back_populates='balance')  # noqa: F821
 	created_at: datetime = Field(
 		default_factory=lambda: datetime.now(UTC),
 		nullable=False,
